+++ src/main/java/com/anz/controller/TransactionController.java
@@ -3,66 +3,84 @@ package com.anz.controller;
 import com.anz.dao.TransactionDAO;
 import com.anz.model.Account;
 import com.anz.model.Transaction;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.URI;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 public class TransactionController {
-    TransactionDAO transactionDAO;
+    private TransactionDAO transactionDAO;
 
     public TransactionController(TransactionDAO transactionDAO) {
         this.transactionDAO = transactionDAO;
     }
 
-    @RequestMapping(value = "/transaction/{accountId}")
-    public ResponseEntity<List<Transaction>> transactions(@PathVariable String accountId) {
+    @GetMapping(value = "/transactions")
+    public ResponseEntity<List<Transaction>> transactions() {
         try {
-            HttpHeaders headers = new HttpHeaders();
-            headers.add("Responded", "TransactionController");
-            return ResponseEntity.accepted().headers(headers).body(transactionDAO.getAll(accountId));
+            return ResponseEntity.ok(transactionDAO.getAll());
         } catch (SQLException e) {
-            return new ResponseEntity<>(new ArrayList<>(), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @PostMapping("/transaction")
-    public ResponseEntity<Transaction> newTransaction (@RequestBody Transaction transaction) throws SQLException {
+    @GetMapping(value = "/transactions/account/{accountId}")
+    public ResponseEntity<List<Transaction>> transactionsFromAccountId(@PathVariable Long accountId) {
         try {
-            transactionDAO.persist(transaction);
-            return ResponseEntity.status(HttpStatus.CREATED).build();
+            return ResponseEntity.ok(transactionDAO.getAll(accountId));
+        } catch (SQLException e) {
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
 
+    @GetMapping("/transactions/transaction/{transactionId}")
+    public ResponseEntity<Transaction> transactionsFromTransactionId(@PathVariable String transactionId) {
+        try {
+            return ResponseEntity.ok(transactionDAO.getTransaction(transactionId));
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    @PostMapping("/transactions")
+    public ResponseEntity<Transaction> newTransaction(@RequestBody Transaction transaction) {
+        try {
+            transactionDAO.create(transaction);
+            return ResponseEntity.created(URI.create("http://localhost:8080/transactions/transaction/" + transaction.getTransactionId())).body(transaction);
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PutMapping("/transactions")
+    public ResponseEntity<Account> updateTransaction(@RequestBody Transaction transaction) {
+        try {
+            transactionDAO.update(transaction);
+            return ResponseEntity.ok().build();
         } catch (SQLException e) {
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @PutMapping("/transaction")
-    public ResponseEntity<Account> updateTransaction(@RequestBody Transaction transaction) throws SQLException {
+    @DeleteMapping("/transactions/{transactionId}")
+    public ResponseEntity<String> deleteTransaction(@PathVariable("transactionId") Long transactionId) {
         try {
-            transactionDAO.update(transaction);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).build();
-
+            Transaction transactionIdToBeDeleted = transactionDAO.getTransaction(String.valueOf(transactionId));
+            if (transactionIdToBeDeleted != null)
+                transactionDAO.delete(transactionId);
+            else
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return ResponseEntity.ok().build();
         } catch (SQLException e) {
             e.printStackTrace();
-            return new ResponseEntity<>( HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    @DeleteMapping("/transactions/{transactionID}")
-    public ResponseEntity<String> deleteTransaction (@PathVariable("transactionId") Long transactionId) throws SQLException {
-        try {
-            transactionDAO.delete(transactionId);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).build();
-        } catch (SQLException e) {
-            e.printStackTrace();
-            return new ResponseEntity<>(new String(), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
